user-service/requirements.txt
requests
datetime
flask
psycopg2
flask_cors
flask-socketio 
# Flask-SocketIO==4.3.1
# python-engineio==3.13.2
# python-socketio==4.6.0
eventlet
uwsgi
gevent
gunicorn
pyJWT